Implement parser.
Expressions can now be tokenized, lexed, and parsed. The evaluated
value is printed into the repl. Order of operations are handled
with a hacky mutation in mparser/src/ast.c[read_ast]. This should
be replaced with a better recursive parsing setup. Rather then
calling read_value() in the right hand side of read_expression()
we should just call read_expression. That would favor right-hand
evaulation of the expressions and entierly remove the need for
this hack.