Remove CMakeLists.
I added it to try and make this compatable with CLion but that did
not work. CLion couldn't see any of my C standard library headers
(somehow?) and I realized the GDB interface it provides would not
work with the piplining system. I'd have to create mock inputs
and outputs for each program. This isn't bad, just annoying.